estruturação e serviço do xlsx funcionando para leitura

+++ src/main/java/br/com/acamargo/api/produto/DadosListagemProdutos.java
@@ -1,6 +1,6 @@
 package br.com.acamargo.api.produto;
 
-public record DadosListagemProdutos(long codigo, String nome, float preco) {
+public record DadosListagemProdutos(int codigo, String nome, float preco) {
 
     public DadosListagemProdutos(Produto produto) {
         this(produto.getCodigo(), produto.getNome(), produto.getPreco());
